+++ src/cs5004/animator/action/Move.java
@@ -20,7 +20,7 @@ public class Move extends AbstractAction {
     this.type = Action.MOVE;
   }
 
-  public IShape getShapeAtTick(int tick, IShape shape) {
+  public IShape getShapeAtTick(double tick, IShape shape) {
     if (tick < 0) {
       throw new IllegalArgumentException("Ticks cannot be negative");
     }
